Adds model classes and file generators
Model classes to represent the pipeline and its configuration have
been added, together with the text generation classes.

As part of this commit, the Behave and Hamcrest testing libraries
have been also added to test the project features.

This additions have not been integrated into the main application,
and this should be done in the future.