Adds support to open help URLs when running in Docker
The "Info" button was not working inside the Docker container, as it
does not have access to the host's browser.

In order to allow this, a workaround was created by using a fifo file
to send the URLs from the container to the host and by watching this
file in the host to open the URLs received though it using the
xdg-open command.