Fixes empty parameter values
When the command of a step was changed, the value of the parameters
was empty. This was caused because the default params values declared
in the Command class was not used by the CommandConfiguration class.

This commit fixes this error and now the CommandConfiguration class
uses the default Command parameter when no parameters are provided to
the constructor.