Adds support for parameter validation
Following the changes in the previous commit, now the GUI supports
parameter validation.

To do so, now a warning is shown when the parameters of a command are
incomplete. This also has been done for the input and output fields.

An additional change done also in this commit is that now config file
is generated even when the pipeline is incomplete or when some
parameter values are missing.