Adds support for different auto-phylo versions
The pipeliner now supports different auto-phylo versions by
dinamically searching for and loading the commands.json file
associated with each version.

By default, the pipeliner will search online for new versions of
auto-phylo and will download automatically their commands.json file,
which will be stored in the ".auto-phylo" directory located in the
user's home.

The commands.json of the first supported version of auto-phylo
(v2.0.0) is included in the source code, as it is not included in the
auto-phylo's source code.

For developing purposes, the pipeliner also supports the use of
custom commands.json files, by using the "-c/--commands" argument.

Automatic search for new versions can be disabled using the
"-d/--disable-update-commands" argument.